fix(repository): cast Price to double in ORDER BY for SQLite compatibility
SQLite does not support expressions of type 'decimal' in ORDER BY clauses.
Cast Price to double to resolve database compatibility issue when sorting
products by price ascending or descending.

Fixes error: 'SQLite does not support expressions of type decimal in ORDER BY clauses'

+++ src/Infrastructure/Repositories/Implements/ProductRepository.cs
@@ -232,8 +232,8 @@ public class ProductRepository : IProductRepository
         // R70: Apply sorting based on SortBy parameter
         query = searchParams.SortBy switch
         {
-            ProductSortOption.PriceAsc => query.OrderBy(p => p.Price),
-            ProductSortOption.PriceDesc => query.OrderByDescending(p => p.Price),
+            ProductSortOption.PriceAsc => query.OrderBy(p => (double)p.Price),
+            ProductSortOption.PriceDesc => query.OrderByDescending(p => (double)p.Price),
             ProductSortOption.NameAsc => query.OrderBy(p => p.Title),
             ProductSortOption.NameDesc => query.OrderByDescending(p => p.Title),
             ProductSortOption.Newest or _ => query.OrderByDescending(p => p.CreatedAt),
